Add logout btn to navbar,
remove from home and account pages
