+++ password-manager/src/pages/Account.js
@@ -1,12 +1,10 @@
 import React, { useState, useEffect } from 'react';
 import { auth } from '../firebase'; // Import Firebase auth
 import UserProfile from '../components/UserProfile';
-import AccountActions from '../components/AccountActions';
 import { useNavigate } from 'react-router-dom';
 
 const Account = () => {
   const [user, setUser] = useState(null);
-  const navigate = useNavigate();
 
   useEffect(() => {
     const fetchUserData = () => {
@@ -18,15 +16,6 @@ const Account = () => {
 
     fetchUserData();
   }, []);
-
-  const handleLogout = async () => {
-    try {
-      await auth.signOut();
-      navigate('/');
-    } catch (error) {
-      console.error('Error logging out:', error);
-    }
-  };
 
   const handleChangePassword = () => {
     // Implement change password functionality
@@ -44,7 +33,14 @@ const Account = () => {
     <div className="min-h-screen flex flex-col items-center bg-background text-text p-6">
       <h2 className="text-4xl font-bold mb-6">Account Page</h2>
       <UserProfile user={user} />
-      <AccountActions handleChangePassword={handleChangePassword} handleLogout={handleLogout} />
+      <div className="mt-4 space-y-2">
+        <button
+          className="w-full py-2 px-4 bg-primary text-white rounded-md hover:bg-primary-dark"
+          onClick={handleChangePassword}
+        >
+          Change Password
+        </button>
+      </div>
     </div>
   );
 };
